(, cast

+++ src/main/java/got/client/gui/GOTGuiAchievements.java
@@ -222,7 +222,7 @@ public class GOTGuiAchievements extends GOTGuiMenuBase {
 			if (i < 0) {
 				i = -1;
 			}
-			currentScroll -= (double) i / j;
+			currentScroll -= (float) ((double) i / j);
 			if (currentScroll < 0.0f) {
 				currentScroll = 0.0f;
 			}
